+++ NEWS.md
@@ -13,6 +13,7 @@ variables do not exist in analysis dataset.
 * `MNG01` uses a `ggtheme` argument to set graphic parameters instead of the now defunct `show_h_grid`, `show_v_grid`
 and `legend_pos` arguments. The `table` arguments now controls the behavior of the table. The arguments `show_n` and
 `show_table` are now defunct.
+* The `precision` argument now control the numbers of digits presented in `DMT01`.
 
 # chevron 0.2.2